simplify NullTracer
