+++ src/trace/BUILD
@@ -5,9 +5,7 @@ cc_library(
     deps = [
         ":attribute",
         ":event",
-        ":null_span",
         ":null_tracer",
-        ":span",
         ":tracer",
     ],
 )
@@ -20,7 +18,6 @@ cc_library(
     include_prefix = "warren/json/trace",
     visibility = ["//visibility:public"],
     deps = [
-        ":null_span",
         ":tracer",
     ],
 )
@@ -29,27 +26,6 @@ cc_library(
     name = "tracer",
     hdrs = [
         "tracer.h",
-    ],
-    include_prefix = "warren/json/trace",
-    visibility = ["//visibility:public"],
-)
-
-cc_library(
-    name = "null_span",
-    hdrs = [
-        "null_span.h",
-    ],
-    include_prefix = "warren/json/trace",
-    visibility = ["//visibility:public"],
-    deps = [
-        ":span",
-    ],
-)
-
-cc_library(
-    name = "span",
-    hdrs = [
-        "span.h",
     ],
     include_prefix = "warren/json/trace",
     visibility = ["//visibility:public"],
