+++ src/trace/tracer.h
@@ -5,7 +5,12 @@
 namespace warren {
 namespace trace {
 
-struct Span;
+struct Span {
+  virtual ~Span() = default;
+  virtual void annotate(std::string_view key, std::string_view value) = 0;
+  virtual void event(std::string_view name) = 0;
+  virtual void end() = 0;
+};
 
 struct Tracer {
   virtual ~Tracer() = default;
